Support SipHash-1-3
c-siphash has been updated to support SipHash-1-3 too. Expose the new
functionality through the Python wrapper.

This introduces the use of the Cython.Tempita template engine to
generate the repetitive parts of the Cython source code. Cython is
already a build requirement, thus this does not introduce any new
dependency.